fix: instantiate sequelize within lambda handler (#385)
* fix: instantiate sequelize within lambda handler

* refactor(bootstrap): expose sequelize and app

Expose sequelize in the object returned by `bootstrap()` so that we can
manipulate sequelize as part of the application lifecycle. This is
needed especially in serverless execution contexts, where we have to
close the pooled connections once we are done with invocation.
