+++ src/server/index.ts
@@ -2,6 +2,6 @@ import bootstrap, { logger } from './bootstrap'
 import config from './config'
 
 const port = config.get('port')
-bootstrap().then((app) =>
+bootstrap().then(({ app }) =>
   app.listen(port, () => logger.info(`Listening on port ${port}`))
 )
